Implemented Accounts Database with Clean Architecture.

+++ app/src/main/java/com/aarves/bluepages/database/AccountDAOImpl.java
@@ -21,7 +21,7 @@ public class AccountDAOImpl implements AccountDAO {
     public void deleteAccount(String username, String passwordHash) {
         AccountDataEntity accountDataEntity = accountDatabaseDAO.getByUsername(username);
 
-        if(accountDataEntity.passwordHash.equals(passwordHash)){
+        if(accountDataEntity != null && accountDataEntity.passwordHash.equals(passwordHash)){
             this.accountDatabaseDAO.delete(accountDataEntity);
         }
     }
@@ -34,7 +34,7 @@ public class AccountDAOImpl implements AccountDAO {
 
     @Override
     public boolean isPasswordMatch(String username, String passwordHash) {
-        AccountDataEntity accountDataEntity = accountDatabaseDAO.getByUsername(username);
-        return accountDataEntity.passwordHash.equals(passwordHash);
+        String accountPasswordHash = accountDatabaseDAO.getPasswordHash(username);
+        return passwordHash.equals(accountPasswordHash);
     }
 }
